A patch from Erwan de FERRIERES "improvement on the calendar screen" (https://issues.apache.org/jira/browse/OFBIZ-2941) - OFBIZ-2941
This is an improvement on the view calendar screen, to use more screenlets and less containers. The hardcoded labels have also been removed.
The modifications are visible when creating or editing an event listed in the calendar. Boxes are better rendered and have a title bar as the other parts of the screen. 

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/workeffort@815479 13f79535-47bb-0310-9956-ffa450edef68