Closes "Move all Groovy scripts in components WEB-INF/actions sub-folders" - https://issues.apache.org/jira/browse/OFBIZ-6981
Currently there are few Groovy scripts not in components WEB-INF/actions sub-folders.
Before moving Groovy scripts as we decided in OFBIZ-6810 we should move the few Groovy scripts not in components WEB-INF/actions sub-folders to these WEB-INF/actions sub-folders. So we will not miss one and it will be a "straightforward" task to move then as we decided in OFBIZ-6810.
Note: most of the Groovy scripts not in components WEB-INF/actions sub-folders are related with events. But not all events related Groovy scripts are not placed components WEB-INF/actions sub-folders, just a very few. There are other cases were Groovy scripts are not in components WEB-INF/actions sub-folders, like ShowProductionRun.groovy and OrderReportPrepare.groovy...





git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/workeffort@1743949 13f79535-47bb-0310-9956-ffa450edef68