[OFBIZ-7095] Fixed console error on Calendar Widget pagination. An exception is being thrown when pagination buttons are clicked on the Calendar widget.
While clicking on the Prev or Next button from the Calendar widget in any view (Day, Week or Month), an exception is thrown due to unparseable date format. The main reason behind this was due to the 'start' parameter in the parameters map. At the time of creating a valid context for "getWorkEffortEventsByPeriod", the start parameter was expected to be in Timestamp format but was actually in String.

Thanks Chandan Khandelwal for reporting the issue and providing patch.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/workeffort@1744856 13f79535-47bb-0310-9956-ffa450edef68