Implemented: Create and use an OWASP PolicyFactory for content sanitization in ContentWorker for Birt Report Builder (OFBIZ-9166)
This is still an incomplete feature because I did not find yet a way to complete
the BIRT_REPORT_BUILDER_USAGE_POLICY and had to bypass the sanitization for this
case, WIP...

This commit has no effect on OFBiz yet.

I need to commit this now to continue to work on the main task: OFBIZ-6919
"New implementation of Birt. Easier user possibility of report creation". 







git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/workeffort@1777451 13f79535-47bb-0310-9956-ffa450edef68